post beta-test changes: - MB-02 handling integrated into File Manager (WIP) - new types of dialogs (including toast) - rom/disk menu checking filepath validity - various bugfixes

+++ arm/utils/file.h
@@ -2,6 +2,7 @@
 #define UTILS_FILE_H_INCLUDED
 
 #include "../system.h"
+#include "cstring.h"
 
 bool FileExists(const char *str);
 bool ReadLine(FIL *file, CString &str);
@@ -10,8 +11,8 @@ bool WriteLine(FIL *file, const char *str);
 
 //---------------------------------------------------------------------------------------
 
-byte read_file(FIL *f, byte *dst, byte sz);
-byte write_file(FIL *f, byte *dst, byte sz);
+int read_file(FIL *f, byte *dst, int sz);
+int write_file(FIL *f, byte *dst, int sz);
 byte read_le_byte(FIL *f, byte *dst);
 byte read_le_word(FIL *f, word *dst);
 byte read_le_dword(FIL *f, dword *dst);
